Tests/LibWeb/Text/expected/focus-events.txt

   focus target=<INPUT id="a">
focusin target=<INPUT id="a">
blur target=<INPUT id="a">
focusout target=<INPUT id="a">
focus target=<INPUT id="b">
focusin target=<INPUT id="b">
